pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "email-validator>=2.2.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.1",
    "openai>=1.93.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "langchain>=0.3.26",
    "faiss-cpu>=1.11.0",
    "python-docx>=1.2.0",
    "werkzeug>=3.1.3",
]
